refactor: 테스트 코드들의 패키지를 구현코드와 동일하게 변경
