+++ app/src/androidTest/java/nextstep/payments/screens/card/update/AddCardScreenTest.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.screens.card.update
 
 import androidx.compose.ui.test.assertIsEnabled
 import androidx.compose.ui.test.assertIsNotEnabled
@@ -6,8 +6,6 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import nextstep.payments.screens.card.state.CardCompanyState
 import nextstep.payments.screens.card.state.CardState
-import nextstep.payments.screens.card.update.AddCardScreen
-import nextstep.payments.screens.card.update.UpdateCardUiState
 import org.junit.Rule
 import org.junit.Test
